Get documents to work with the new structure
