+++ src/search_analysis/tools/comparison/__init__.py
@@ -1,0 +1,1 @@
+from .comparison_tool import ComparisonTool
