+++ src/search_analysis/tools/elasticsearch/__init__.py
@@ -1,0 +1,1 @@
+from .es_evaluation_object import ESEvaluationObject
